Improved import data. Now returns list of pandas DataFrames for easier processing later on. Added interpolate inital upload.

+++ import_data.py
@@ -1,33 +1,40 @@
 import requests
 import logging
+import pandas
 import numpy
 import json
 import os
+import io
 
 
 def import_data(signalFrom="", signalTo="", signalsFile='./testsignals.json', saveDirectory="./.localData", saveLocal=True, override=False):
     """Fetches data via API and saves locally or returns list of strings"""
-    logger = logging.getLogger(__name__)
-    # Create appropriate directories
-    try:
-        os.listdir(saveDirectory)
-    except FileNotFoundError:
-        logger.info(f"Directory not found: \"{saveDirectory}\"")
-        os.makedirs(saveDirectory)
-        logger.info(f"Directory created: \"{saveDirectory}\"")
     
-    # Override
-    if override:
-        for file in os.listdir(saveDirectory):
-            os.remove(os.path.join(saveDirectory, file))
-            logger.debug(f"Removed all files in directory: \"{saveDirectory}\"")
-    elif not override and os.listdir(saveDirectory):
-        logger.info(f"Files already generated. Pass parameter \"override=True\" to override.")
-        return
+    # Initalize logger
+    logger = logging.getLogger(__name__)
+    
+    if saveLocal:
+        print(saveLocal)
+        # Create appropriate directories
+        try:
+            os.listdir(saveDirectory)
+        except FileNotFoundError:
+            logger.info(f"Directory not found: \"{saveDirectory}\"")
+            os.makedirs(saveDirectory)
+            logger.info(f"Directory created: \"{saveDirectory}\"")
+        # Override
+        if override:
+            for file in os.listdir(saveDirectory):
+                os.remove(os.path.join(saveDirectory, file))
+                logger.debug(f"Removed all files in directory: \"{saveDirectory}\"")
+        elif not override and os.listdir(saveDirectory):
+            logger.info(f"Files already generated. Pass parameter \"override=True\" to override.")
+            return
 
     # Variables
     if not saveLocal:
         returnData = numpy.array([])
+        print('here1')
 
     # Read signals.json
     try:
@@ -43,7 +50,7 @@ def import_data(signalFrom="", signalTo="", signalsFile='./testsignals.json', sa
 
     # Save or return signal data
     for signal in signalsData:
-        csvData = ""
+        csvData = "timestamp, value\n"
         url=f'https://signalapi.bmesh.io/api/sevents/entity/{signal}'
         # "from" and "to" parameter URL constructing logic 
         if signalFrom or signalTo:
@@ -67,7 +74,12 @@ def import_data(signalFrom="", signalTo="", signalsFile='./testsignals.json', sa
                 file.write(csvData)
                 logger.info(f"Wrote file: \"{signal}.csv\"")
         else:
-            returnData = numpy.append(returnData, csvData)
-    
+            # returnData = numpy.append(returnData, csvData)
+            # print(returnData)
+            # returnDataDF = pandas.DataFrame(data=returnData)
+            # print('here2')
+            returnDataDF = pandas.read_csv(io.StringIO(csvData), sep=',')
+            returnData = numpy.append(returnData, returnDataDF)
+
     if not saveLocal:
         return returnData
